[♻️refactor]: MyComponent 파일 경로 변경
- src/components로 수정
- src/app/component 디렉토리는 URL 상 /component 페이지를 나타내기 때문에
  component 들은 src/components에 담음
