+++ src/components/MyComponent/MyComponent.style.ts
